Change namespaces
We were inconsistent in whether we used http://obdurodon.org/hoax
or http://www.obdurodon.org/hoax as the main project namespace
(bound to the prefix "hoax:"). We standardize on
http://www.obdurodon.org/hoax.

Separately from our previous inconsistency about "www.", the old
model namespace was http://www.obdurodon.org/model, which did not
document a connection to the hoax project. We change it to
http://www.obdurodon.org/hoax/model. We continue to use the prefix
"m:".